requirements.txt
asgiref==3.5.2
black==22.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
defusedxml==0.7.1
Django==4.0.5
django-cors-headers==3.12.0
django-environ==0.10.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
fonttools==4.39.4
fpdf2==2.7.4
freetype-py==2.3.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.5.0
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.6
pycairo==1.23.0
PyJWT==2.7.0
pytz==2022.1
reportlab==4.0.0
requests==2.30.0
rlPyCairo==0.2.0
sentry-sdk==1.22.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.11
